Refactor ResourceTypeHelper and update callers
I've replaced getNewOrBreakingResourceTypeNames() with a more
generic `isCompatible(resType, fhirVersion1, fhirVersion2)` method.

To implement it, I introduced new members to the ResourceTypeName enums
so that we can compute which resource types are associated with which
FHIR versions.

The compatibility knowledge is still encapsulated in
ResourceTypeHelper, but now we more cleanly distinguish between
R4B_ONLY_RESOURCE_TYPES and BACKWARD_BREAKING_R4B_ENUMS (previously they
were both together in the R4B_ONLY_RESOURCE_TYPES list). And now we have
a way to get the inverse (valid 4.3 resources that maybe not be valid in
4.0), although this isn't currently used by the server.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>